requirements.txt
streamlit
openai
pandas
openpyxl
python-dotenv

